Remove clientMutationId pattern
Client mutation IDs are no longer required to uniquely identify
mutation requests as of the latest Relay release.
Passing a single input object as the mutation data is also no longer
necessary.

This updates the examples to remove these old pattern to show the latest
patterns.

NOTE: continuing to pass a clientMutationId or using the "input object
pattern" for mutations will continue to work.